refactor: group typography related components
